moved test to DurableTaskEndToEndTests and cleaned up test code
