+++ test/Common/DurableClientBaseTests.cs
@@ -7,8 +7,6 @@ using System.Linq;
 using System.Net.Http;
 using System.Threading.Tasks;
 using DurableTask.Core;
-using FluentAssertions;
-using Microsoft.Azure.WebJobs.Host.TestCommon;
 #if !FUNCTIONS_V1
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
@@ -28,14 +26,7 @@ namespace Microsoft.Azure.WebJobs.Extensions.DurableTask.Tests
 {
     public class DurableClientBaseTests
     {
-        private readonly ITestOutputHelper output;
-        private readonly TestLoggerProvider loggerProvider;
-
-        public DurableClientBaseTests(ITestOutputHelper output)
-        {
-            this.output = output;
-            this.loggerProvider = new TestLoggerProvider(output);
-        }
+        public DurableClientBaseTests() { }
 
         [Theory]
         [Trait("Category", PlatformSpecificHelpers.TestCategory)]
@@ -221,30 +212,6 @@ namespace Microsoft.Azure.WebJobs.Extensions.DurableTask.Tests
 
             var durableOrchestrationClient = this.GetDurableClient(orchestrationServiceClientMock.Object);
             Assert.ThrowsAny<InvalidOperationException>(() => durableOrchestrationClient.CreateHttpManagementPayload("testInstanceId"));
-        }
-
-        /// <summary>
-        /// End to end test that ensures that DurableClientFactory is set up correctly
-        /// (i.e. the correct services are injected through dependency injection
-        /// and AzureStorageDurabilityProvider is created).
-        /// </summary>
-        [Fact]
-        [Trait("Category", PlatformSpecificHelpers.TestCategory)]
-        public async Task DurableClient_AzureStorage_SuccessfulSetup()
-        {
-            string orchestratorName = nameof(TestOrchestrations.SayHelloInline);
-            using (ITestHost host = TestHelpers.GetJobHost(
-                loggerProvider: this.loggerProvider,
-                testName: nameof(this.DurableClient_AzureStorage_SuccessfulSetup),
-                enableExtendedSessions: false,
-                storageProviderType: "azure_storage",
-                addDurableClientFactory: true))
-            {
-                await host.StartAsync();
-                var client = await host.StartOrchestratorAsync(orchestratorName, input: "World", this.output);
-                var status = await client.WaitForCompletionAsync(this.output);
-                await host.StopAsync();
-            }
         }
 
         private IDurableOrchestrationClient GetDurableClient(IOrchestrationServiceClient orchestrationServiceClientMockObject)
